pyproject.toml
[tool.poetry]
name = "pythonplaywright"
version = "0.1.0"
description = "Test framework using Python, Pytest and Playwright"
authors = ["Jaroslav Janosik <jarajz.janosik@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.5"
playwright = "^1.52.0"
pytest-playwright = "^0.7.0"
pytest = "^8.4.1"
pytest-xdist = "^3.7.0"
pytest-bdd = "^8.1.0"
pytest-html = "^4.1.1"
google-api-python-client = "^2.173.0"
google-auth-oauthlib = "^1.1.0"
poetry-core = "^2.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
